engine: extract esm/constellation_filter.js
