+++ engine/esm/index.js
@@ -174,6 +174,7 @@ export {
 } from "./interfaces.js";
 export { Annotation, Circle, Poly, PolyLine } from "./annotation.js";
 export { SolarSystemObjects, InterpolationType, CameraParameters } from "./camera_parameters.js";
+export { ConstellationFilter } from "./constellation_filter.js";
 export { FitsProperties, ScaleTypes } from "./fits_properties.js";
 export { Star, Galaxy } from "./star.js";
 export { UiTools } from "./ui_tools.js";
@@ -243,7 +244,6 @@ export {
     Constellations,
     Lineset,
     Linepoint,
-    ConstellationFilter,
     Folder,
     FolderBrowser,
     FolderUp,
